使用TableColumn，使代码更具React-style
